WIP like button experiment
This does not currently work but was an experiment to add a like button
to the mention section. The idea was that people could click the like
without needing to log in or have a website.

Doing this via comentpara.de was clunky. So I was looking into creating
an edge function that would return a valid webmention source for
webmention.io.